refactor: create repository, service, util directory
