+++ src/main/java/lotto/domain/repository/WinningRanks.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.repository;
 
 import java.util.List;
 import java.util.Objects;
