+++ src/main/java/lotto/domain/service/NumberSelectionStrategy.java
@@ -1,4 +1,6 @@
-package lotto.domain;
+package lotto.domain.service;
+
+import lotto.domain.repository.Numbers;
 
 @FunctionalInterface
 public interface NumberSelectionStrategy {
