+++ src/main/java/lotto/domain/util/MoneyLottoMapper.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.util;
 
 public class MoneyLottoMapper {
