+++ src/test/java/lotto/domain/LottosTest.java
@@ -1,5 +1,7 @@
 package lotto.domain;
 
+import lotto.domain.repository.Lottos;
+import lotto.domain.service.NumberSelectionStrategyImpl;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
