+++ src/test/java/lotto/domain/MoneyLottoMapperTest.java
@@ -1,5 +1,6 @@
 package lotto.domain;
 
+import lotto.domain.util.MoneyLottoMapper;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
